# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  httputil.Error400Response:
    properties:
      error:
        example: invalid request body
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  httputil.Error500Response:
    properties:
      error:
        example: internal server error
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  staff.CreateStaffInput:
    properties:
      branch_id:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  staff.Staff:
    properties:
      branch_id:
        type: string
      business_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        type: string
    type: object
  staff.UpdateStaffInput:
    properties:
      branch_id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  table.CreateTableInput:
    properties:
      branch_id:
        type: string
      name:
        type: string
      qr_code:
        type: string
    type: object
  table.Table:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      qr_code:
        type: string
      status:
        type: string
    type: object
  table.UpdateTableInput:
    properties:
      name:
        type: string
      qr_code:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant POS SaaS backend
  title: Bayar.in API
  version: "1.0"
paths:
...
  /staff:
    get:
      description: Returns all staff members for the authenticated owner's business
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/staff.Staff'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: List staff
      tags:
      - Staff
    post:
      consumes:
      - application/json
      description: Create a new cashier account. Role is always set to "cashier".
        branch_id is required.
      parameters:
      - description: Staff payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/staff.CreateStaffInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/staff.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Create staff (cashier)
      tags:
      - Staff
  /staff/{id}:
    patch:
      consumes:
      - application/json
      description: Update a staff member's name, active status, or assigned branch.
        Cannot deactivate owner.
      parameters:
      - description: Staff UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/staff.UpdateStaffInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update staff
      tags:
      - Staff
  /tables:
    get:
      description: Returns all tables. Filter by branch_id query param (optional).
      parameters:
      - description: Branch UUID to filter tables
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/table.Table'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: List tables
      tags:
      - Tables
    post:
      consumes:
      - application/json
      description: Create a new table under a branch. The branch must belong to the
        authenticated owner's business.
      parameters:
      - description: Table payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/table.CreateTableInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/table.Table'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Create table
      tags:
      - Tables
  /tables/{id}:
    patch:
      consumes:
      - application/json
      description: Partially update a table's name, status ("available"|"occupied"),
        or QR code
      parameters:
      - description: Table UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/table.UpdateTableInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/table.Table'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update table
      tags:
      - Tables
securityDefinitions:
  BearerAuth:
    description: 'Paste your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"